requirements.txt
bokeh==2.4.2
numpy==1.22.1
pandas==1.3.5
panel==0.12.6
PyYAML==6.0
scikit_bio==0.5.6
scipy==1.7.3
